New ConfigGUI
The ConfigGUI has been completely revamped. It looks much better, with the new Windows 8/10 Metro design.

Key notes:
- The new config keeps the same categories, with the inclusion of the "Coming Soon" CPE (Classic Protocal Extension) Category.
- To access each of the categories, you just need to hit the corresponding button. A window will appear with the familliar layout for the category that you know and recognize.

In order to make this possible, I had to seperate each of the config sections into different Forms of their own, and set all their members to "internal" because of the amount of input handlers there were and modifications in the MainForm class.