More work on custom blocks
Custom blocks loaded from same extension (gcblock) but are a Zip archive.

Zip archive has 4 mandatory files
- Bottom Texture (png)
- Top Texture (png)
- Side Texture (png)
- block.json

Even if one of the textures is the same as another, they still need to be in the archive. GemsCraft determines if these images are the same and will handle it based on that.

Also to note is the simplified ID checkers. Previously it was believed that the extra brick textures and the gold texture on terrain.png were used by the client, but that is not the case. So therefore I was able to simplify how I checked ID's.

Next step is to setup uploading to webserver (which is started as can be seen). In order to create unqiue url's for the classicube client to load, the file will be uploaded as an md5 hash string of the server IP and port together.