Am I a Pretty Girl?
Code improvements, also updated the updater...
And then the name here actually has meaning. The last thing I was working on was the GUI, giving it the beautiful background images it so deserves, and also updating the icons to the new icon. I'm tired though, so I'll finish the progress I made tomorrow