Texture Packs
Server owners can now select a texture pack they have on their file system from the World Section of the Configuration (a terrain.png or a zip)

Custom Blocks are also automatically added and will carry out like they would with the default texture pack.

This also allows for the use of custom blocks and to maintain the animations of animated blocks.