Configurable Plugins
Users can now decide if they want to enable or disable plugins.
Users can also now configure plugins based on properties in the Init.cs class in the plugin.

And in order to keep consistency and functionality stable, plugins must adhere to the version of GemsCraft being used. If the version does not match, GemsCraft will refuse to load the plugin.